pulling out index tests
