+++ .github/workflows/parallel-tests.yaml
@@ -29,12 +29,13 @@ jobs:
           name: build-artifacts
           path: build/
           retention-days: 1
+
   test-bundle:
     runs-on: ubuntu-latest
     needs: build
     strategy:
       matrix:
-        test: [bundle, deploy-flags]
+        test: [bundle, deploy-flags, inspect]
     steps:
       - name: Checkout
         uses: actions/checkout@692973e3d937129bcbf40652eb9f2f61becf3332 # v4.1.7
